Proyecto2/docker/API/app/requirements.txt
flask==2.2.3
azure-identity==1.12.0
azure-storage-blob==12.16.0
Flask-Cors==3.0.10
requests==2.28.2